+++ backend/api/urls.py
@@ -8,6 +8,6 @@ urlpatterns = [
     path("api/", include(router.urls)),
     path("api/admin", admin.site.urls),  # do i want this ?
     path("api/auth/", include("modules.accounts.urls")),
-    path("api/<str:language_code>/", include("cards.urls")),
+    path("api/<str:language_code>/", include("modules.cards.urls")),
     path("api/", include("modules.card_collections.urls")),
 ]
