+++ backend/modules/cards/apps.py
@@ -3,4 +3,4 @@ from django.apps import AppConfig
 
 class CardsConfig(AppConfig):
     default_auto_field = "django.db.models.BigAutoField"
-    name = "cards"
+    name = "modules.cards"
