+++ backend/modules/cards/management/commands/download_cards_images_fr.py
