+++ backend/modules/cards/management/commands/download_colors_images.py
