+++ backend/modules/cards/management/commands/download_pokedex_data.py
