+++ backend/modules/cards/management/commands/download_rarity_images.py
