+++ backend/modules/cards/management/commands/languages_seeder.py
@@ -1,5 +1,5 @@
 from django.core.management.base import BaseCommand
-from cards.models import (
+from modules.cards.models import (
     Language,
 )
