+++ backend/modules/cards/management/commands/rarities_seeder.py
@@ -1,6 +1,6 @@
 import json
 from django.core.management.base import BaseCommand
-from cards.models import (
+from modules.cards.models import (
     Language,
     Rarity,
     RarityTranslation,
