+++ backend/modules/cards/utils.py
